+++ file/test/file_endpoint_test.exs
@@ -1,6 +1,8 @@
 defmodule Membrane.RTC.FileEndpointTest do
   use ExUnit.Case
 
+  alias ExSDP.Attribute.FMTP
+
   alias Membrane.RTC.Engine
   alias Membrane.RTC.Engine.Endpoint
   alias Membrane.RTC.Engine.Message
@@ -26,8 +28,6 @@ defmodule Membrane.RTC.FileEndpointTest do
 
   @source_endpoint_id "source-endpoint-id"
   @sink_endpoint_id "sink-endpoint-id"
-  @track_id "test-track-id"
-  @stream_id "test-stream"
 
   describe "File Endpoint test" do
     @tag :tmp_dir
@@ -38,7 +38,7 @@ defmodule Membrane.RTC.FileEndpointTest do
       file_name = "audio.opus"
       file_path = Path.join(@fixtures_dir, file_name)
 
-      output_file = Path.join([tmp_dir, "#{@track_id}.opus"])
+      output_file = Path.join([tmp_dir, "out_audio.opus"])
 
       :ok =
         Engine.add_endpoint(
@@ -53,10 +53,7 @@ defmodule Membrane.RTC.FileEndpointTest do
       file_endpoint =
         create_audio_file_endpoint(
           rtc_engine,
-          file_path,
-          @stream_id,
-          @source_endpoint_id,
-          @track_id
+          file_path
         )
 
       :ok = Engine.add_endpoint(rtc_engine, file_endpoint, id: @source_endpoint_id)
@@ -98,7 +95,7 @@ defmodule Membrane.RTC.FileEndpointTest do
       file_name = "video.h264"
       file_path = Path.join(@fixtures_dir, file_name)
 
-      output_file = Path.join([tmp_dir, "#{@track_id}.h264"])
+      output_file = Path.join([tmp_dir, "out_video.h264"])
 
       :ok =
         Engine.add_endpoint(
@@ -113,10 +110,7 @@ defmodule Membrane.RTC.FileEndpointTest do
       file_endpoint =
         create_video_file_endpoint(
           rtc_engine,
-          file_path,
-          @stream_id,
-          @source_endpoint_id,
-          @track_id
+          file_path
         )
 
       :ok = Engine.add_endpoint(rtc_engine, file_endpoint, id: @source_endpoint_id)
@@ -153,40 +147,29 @@ defmodule Membrane.RTC.FileEndpointTest do
 
   defp create_video_file_endpoint(
          rtc_engine,
-         video_file_path,
-         stream_id,
-         video_source_endpoint_id,
-         video_track_id
+         video_file_path
        ) do
-    video_track =
-      Engine.Track.new(
-        :video,
-        stream_id,
-        video_source_endpoint_id,
-        :H264,
-        90_000,
-        %ExSDP.Attribute.FMTP{
-          pt: 96
-        },
-        id: video_track_id,
-        framerate: {60, 1}
-      )
+    video_track_config = %Endpoint.File.TrackConfig{
+      type: :video,
+      encoding: :H264,
+      clock_rate: 90_000,
+      fmtp: %FMTP{
+        pt: 96
+      },
+      opts: [framerate: {60, 1}]
+    }
 
     %Endpoint.File{
       rtc_engine: rtc_engine,
       file_path: video_file_path,
-      track: video_track,
-      ssrc: 1234,
+      track_config: video_track_config,
       payload_type: 96
     }
   end
 
   defp create_audio_file_endpoint(
          rtc_engine,
-         audio_file_path,
-         stream_id,
-         audio_source_endpoint_id,
-         audio_track_id
+         audio_file_path
        ) do
     ext = String.split(audio_file_path, ".") |> List.last()
 
@@ -196,22 +179,19 @@ defmodule Membrane.RTC.FileEndpointTest do
         "opus" -> :OPUS
       end
 
-    audio_track =
-      Engine.Track.new(
-        :audio,
-        stream_id,
-        audio_source_endpoint_id,
-        encoding,
-        48_000,
-        nil,
-        id: audio_track_id
-      )
+    audio_track_config = %Endpoint.File.TrackConfig{
+      type: :audio,
+      encoding: encoding,
+      clock_rate: 48_000,
+      fmtp: %FMTP{
+        pt: 108
+      }
+    }
 
     %Endpoint.File{
       rtc_engine: rtc_engine,
       file_path: audio_file_path,
-      track: audio_track,
-      ssrc: 2345,
+      track_config: audio_track_config,
       payload_type: 108
     }
   end