+++ file/test/support/file_sink_endpoint.ex
@@ -6,10 +6,8 @@ defmodule Membrane.RTC.Engine.Support.Sink do
 
   require Membrane.Logger
 
-  alias Membrane.ChildrenSpec.Builder
   alias Membrane.File.Sink, as: FileSink
   alias Membrane.RTC.Engine
-  alias Membrane.RTC.Engine.Endpoint.File, as: FileEndpoint
   alias Membrane.RTC.Engine.Endpoint.WebRTC.TrackReceiver
   alias Membrane.RTC.Engine.Track