+++ file/CHANGELOG.md
@@ -2,5 +2,7 @@
 
 ## 0.2.0-dev
 
+* Added track config [#329](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/329)
+
 ## 0.1.0
 * Initial release [#323](https://github.com/jellyfish-dev/membrane_rtc_engine/pull/323)